Support content import, add confirmations, and small bugfixes

+++ qs-init.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 ##
-## USAGE: ./qs-init.sh --client UNIQUE_SLUG --git-repo GIT_REMOTE [--up]
+## USAGE: ./qs-init.sh --client UNIQUE_SLUG --git-repo GIT_REMOTE [--wxr WXR_TO_IMPORT] [--theme DIRECTORY_NAME] [--up]
 ## Include `--up` when setting up a fresh instance of this VM.
 ##
 
@@ -12,6 +12,8 @@ printf '\nPreparing to initialize the VIP Go Quickstart environment...\n\n'
 printf '1) Parsing arguments...\n\n'
 client=''
 client_git_repo=''
+theme=0
+wxr=0
 needs_to_up=0
 
 while :; do
@@ -38,6 +40,25 @@ while :; do
             fi
             ;;
 
+        --wxr)
+            if [ -n "$2" ]; then
+                rm data/import.xml
+                cp "$2" data/import.xml
+                printf 'CONTENT: Using specified WXR to initialize environment.\n\n'
+                wxr=1
+                shift 2
+                continue
+            fi
+            ;;
+
+         --theme)
+            if [ -n "$2" ]; then
+                theme=$2
+                shift 2
+                continue
+            fi
+            ;;
+
         --up)
             needs_to_up=1
             ;;
@@ -59,21 +80,41 @@ if [ -z "$client_git_repo" ]; then
     exit 1
 fi
 
+if [ "$wxr" = 0 ]; then
+    git checkout -- data/import.xml
+    printf 'CONTENT: Using default WXR to initialize environment.\n\n'
+fi
+
 export VIP_GO_CLIENT=$client
 export VIP_GO_CLIENT_GIT=$client_git_repo
+export VIP_GO_THEME=$theme
 
-printf "CLIENT: %s\nGIT REMOTE: %s\nVAGRANT UP REQUESTED: %d\n\n" "$client" "$client_git_repo" "$needs_to_up"
+printf "CLIENT: %s\nGIT REMOTE: %s\nTHEME: %s\nVAGRANT UP REQUESTED: %d\n\n" "$client" "$client_git_repo" "$theme" "$needs_to_up"
+
+# Require confirmation before destructive actions that follow
+read -p "Continue (y/n)? " CONT
+if [ "$CONT" != "y" ]; then
+  printf 'Aborting at your request.\n';
+  exit 0;
+fi
 
 # Load client's custom code into VM's NFS shares
 # Can't simply delete `go-client-repo` and replace it, as removing the synced folders sends Vagrant into a tizzy
 printf '2) Loading new client code...\n\n'
 rm -rf go-client-repo-new/
+rm -rf go-client-repo/.git/
 rm -rf go-client-repo/languages/*
 rm -rf go-client-repo/plugins/*
 rm -rf go-client-repo/themes/*
 git clone "$client_git_repo" go-client-repo-new
-cp -r go-client-repo-new/* go-client-repo
+cp -r go-client-repo-new/. go-client-repo
 rm -rf go-client-repo-new/
+
+# Ensure theme directory exists, otherwise exit with an error
+if [ "$theme" != 0 ] && [ ! -d "./go-client-repo/themes/$theme" ]; then
+    printf '\n\nERROR: Theme directory "%s" not found. Please check your entry, or omit the "--theme" argument, and try again.\n' "$theme" >&2
+    exit 1
+fi
 
 # Ensure optional languages directory exists
 # Vagrant doesn't respond well if the directory exists sometimes but not always
@@ -82,7 +123,7 @@ if [ ! -d "./go-client-repo/languages/" ]; then
 fi
 
 # Provision the VM
-if [ $needs_to_up == 1 ]; then
+if [ $needs_to_up = 1 ]; then
     printf '\n3) Starting to provision VM. vagrant up can take some time...\n\n'
     vagrant up
 else
